@@ -1,127 +1,136 @@
-var fs = require('fs');
+import fs from "fs";
+import path from 'path';
+import { fileURLToPath } from 'url';
 
-var hasOwn = Object.prototype.hasOwnProperty;
-var toStr = Object.prototype.toString;
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
-var isArray = function isArray(arr) {
-	if (typeof Array.isArray === 'function') {
-		return Array.isArray(arr);
-	}
+const hasOwn = Object.prototype.hasOwnProperty;
+const toStr = Object.prototype.toString;
 
-	return toStr.call(arr) === '[object Array]';
+const isArray = function isArray(arr) {
+    if (typeof Array.isArray === 'function') {
+        return Array.isArray(arr);
+    }
+
+    return toStr.call(arr) === '[object Array]';
 };
 
-var isPlainObject = function isPlainObject(obj) {
-	if (!obj || toStr.call(obj) !== '[object Object]') {
-		return false;
-	}
+const isPlainObject = function isPlainObject(obj) {
+    if (!obj || toStr.call(obj) !== '[object Object]') {
+        return false;
+    }
 
-	var hasOwnConstructor = hasOwn.call(obj, 'constructor');
-	var hasIsPrototypeOf = obj.constructor && obj.constructor.prototype && hasOwn.call(obj.constructor.prototype, 'isPrototypeOf');
-	// Not own constructor property must be Object
-	if (obj.constructor && !hasOwnConstructor && !hasIsPrototypeOf) {
-		return false;
-	}
+    const hasOwnConstructor = hasOwn.call(obj, 'constructor');
+    const hasIsPrototypeOf = obj.constructor && obj.constructor.prototype && hasOwn.call(obj.constructor.prototype, 'isPrototypeOf');
+    // Not own constructor property must be Object
+    if (obj.constructor && !hasOwnConstructor && !hasIsPrototypeOf) {
+        return false;
+    }
 
-	// Own properties are enumerated firstly, so to speed up,
-	// if last one is own, then all properties are own.
-	var key;
-	for (key in obj) { /**/ }
+    // Own properties are enumerated firstly, so to speed up,
+    // if last one is own, then all properties are own.
+    let key;
+    for (key in obj) { /**/
+    }
 
-	return typeof key === 'undefined' || hasOwn.call(obj, key);
+    return typeof key === 'undefined' || hasOwn.call(obj, key);
 };
 
-function extend() {
-	var options, name, src, copy, copyIsArray, clone;
-	var target = arguments[0];
-	var i = 1;
-	var length = arguments.length;
-	var deep = false;
+export function extend() {
+    let options, name, src, copy, copyIsArray, clone;
+    let target = arguments[0];
+    let i = 1;
+    let deep = false;
 
-	// Handle a deep copy situation
-	if (typeof target === 'boolean') {
-		deep = target;
-		target = arguments[1] || {};
-		// skip the boolean and the target
-		i = 2;
-	} else if ((typeof target !== 'object' && typeof target !== 'function') || target == null) {
-		target = {};
-	}
+    // Handle a deep copy situation
+    if (typeof target === 'boolean') {
+        deep = target;
+        target = arguments[1] || {};
+        // skip the boolean and the target
+        i = 2;
+    } else if ((typeof target !== 'object' && typeof target !== 'function') || target == null) {
+        target = {};
+    }
 
-	for (; i < length; ++i) {
-		options = arguments[i];
-		// Only deal with non-null/undefined values
-		if (options != null) {
-			// Extend the base object
-			for (name in options) {
-				src = target[name];
-				copy = options[name];
+    for (; i < arguments.length; ++i) {
+        options = arguments[i];
+        // Only deal with non-null/undefined values
+        if (options != null) {
+            // Extend the base object
+            for (name in options) {
+                src = target[name];
+                copy = options[name];
 
-				// Prevent never-ending loop
-				if (target !== copy) {
-					// Recurse if we're merging plain objects or arrays
-					if (deep && copy && (isPlainObject(copy) || (copyIsArray = isArray(copy)))) {
-						if (copyIsArray) {
-							copyIsArray = false;
-							clone = src && isArray(src) ? src : [];
-						} else {
-							clone = src && isPlainObject(src) ? src : {};
-						}
+                // Prevent never-ending loop
+                if (target !== copy) {
+                    // Recurse if we're merging plain objects or arrays
+                    if (deep && copy && (isPlainObject(copy) || (copyIsArray = isArray(copy)))) {
+                        if (copyIsArray) {
+                            copyIsArray = false;
+                            clone = src && isArray(src) ? src : [];
+                        } else {
+                            clone = src && isPlainObject(src) ? src : {};
+                        }
 
-						// Never move original objects, clone them
-						target[name] = extend(deep, clone, copy);
+                        // Never move original objects, clone them
+                        target[name] = extend(deep, clone, copy);
 
-					// Don't bring in undefined values
-					} else if (typeof copy !== 'undefined') {
-						target[name] = copy;
-					}
-				}
-			}
-		}
-	}
+                        // Don't bring in undefined values
+                    } else if (typeof copy !== 'undefined') {
+                        target[name] = copy;
+                    }
+                }
+            }
+        }
+    }
 
-	// Return the modified object
-	return target;
+    // Return the modified object
+    return target;
 }
 
-function settings(warnuser, path) {
-    this.val = null;
-    var realpath = null, defaultsettings = {};
-    var defrealpath = fs.realpathSync(__dirname + '/../settings.js');
-    if (fs.existsSync(defrealpath)) {
-        defaultsettings = require(defrealpath);
-    }
-    if (path) {
-        if (fs.existsSync(path)) {
-            realpath = fs.realpathSync(path);
-            this.val = require(realpath);
-            console.log('SETTINGS: Using ' + realpath);
+export class Settings {
+    val = null;
+
+    async init(warnuser, path) {
+        let realPath = null, defaultSettings = {};
+        const defaultSettingsRealPath = fs.realpathSync(__dirname + '/../settings.cjs');
+
+        if (fs.existsSync(defaultSettingsRealPath)) {
+            defaultSettings = (await import(defaultSettingsRealPath))?.default;
         }
-    } else if (fs.existsSync(__dirname + '/../settings-user.js')) {
-        realpath = fs.realpathSync(__dirname + '/../settings-user.js');
-        this.val = require(realpath);
-        console.log('SETTINGS: Using ' + realpath);
-    }
-    if (this.val === null) {
-        this.val = defaultsettings;
-        if (warnuser === true) {
-            console.log('WARNING: You should create a custom settings file -> cp settings.js settings-user.js');
-            console.log('SETTINGS: Using ' + defrealpath);
+        if (path) {
+            if (fs.existsSync(path)) {
+                realPath = fs.realpathSync(path);
+                this.val = (await import(realPath)).default;
+                console.log('SETTINGS: Using ' + realPath);
+            }
+        } else {
+            if (fs.existsSync(__dirname + '/../settings-user.js')) {
+                fs.renameSync(__dirname + '/../settings-user.js', __dirname + '/../settings-user.cjs');
+            }
+            if (fs.existsSync(__dirname + '/../settings-user.cjs')) {
+                realPath = fs.realpathSync(__dirname + '/../settings-user.cjs');
+                this.val = (await import(realPath)).default;
+                console.log('SETTINGS: Using ' + realPath);
+            }
         }
-    } else {
-        this.val = extend(true, {}, defaultsettings, this.val);
+        if (this.val === null) {
+            this.val = defaultSettings;
+            if (warnuser === true) {
+                console.log('WARNING: You should create a custom settings file -> cp settings.cjs settings-user.cjs');
+                console.log('SETTINGS: Using ' + defaultSettingsRealPath);
+            }
+        } else {
+            this.val = extend(true, {}, defaultSettings, this.val);
+        }
+
+        if (!this.val.default.theme) {
+            this.val.default.theme = 'default';
+        }
     }
-    
-    if (!this.val.default.theme) {
-        this.val.default.theme = 'default';
-    }
-    
-    return this;
 }
 
-settings.prototype.prefix = function(s) {
-	return this.val.prefixpath + s;
+Settings.prototype.prefix = function (s) {
+    return this.val.prefixpath + s;
 };
-
-exports.settings = settings;
-exports.extend = extend;
