@@ -1,41 +1,41 @@
-const WebSocket = require('ws');
-const net = require('net');
+import { WebSocketServer } from "ws";
+import net from "net";
 
 
 function createSocket(host, port, callback) {
-    var socket = net.connect({
+    const socket = net.connect({
         host: host,
         port: port
     }, function (err) {
         if (err) return callback(err);
     });
-    
+
     socket.on('error', function (err) {
         callback(err);
     });
-    
+
     return socket;
 }
 
-module.exports = function(server, settings) {
-    const wss = new WebSocket.Server({
+export function WS(server, settings) {
+    const wss = new WebSocketServer({
         server: server
     });
-    
+
     wss.on('connection', function connection(ws, req) {
         ws.once('message', function incoming(targetInfo) {
             // First message is the information about the target server
-            var { server, port } = JSON.parse(targetInfo);
-            
+            let { server, port } = JSON.parse(targetInfo);
+
             if (settings.val.forcedefault) {
                 server = settings.val.default.host;
                 port = settings.val.default.port;
             }
-            
+
             const socket = createSocket(server, port, (err) => {
                 if (err) return ws.send(err.toString());
             });
-            
+
             ws.on('message', function (data) {
                 socket.write(data, 'binary', function () {
                     // console.log('Sent: ', (flags.buffer || chunk).toString());
@@ -56,4 +56,4 @@ module.exports = function(server, settings) {
             });
         });
     });
-};
+}
